fix: Blackwell sm_120 compatibility (bf16 defaults)
Override fp16 global defaults to bf16 in devices.py. FP16 causes
silent NaN propagation on Blackwell consumer GPUs (RTX 5090, sm_120).
The AE already blocks fp16 (raises RuntimeError); this patch aligns
the global dtype defaults.

SDPA backend dispatch is already safe — default backend=None enables
all backends (FLASH_ATTENTION, EFFICIENT_ATTENTION, MATH).

+++ SUPIR/utils/devices.py
@@ -66,9 +66,9 @@ enable_tf32()
 
 cpu = torch.device("cpu")
 device = device_interrogate = device_gfpgan = device_esrgan = device_codeformer = torch.device("cuda")
-dtype = torch.float16
-dtype_vae = torch.float16
-dtype_unet = torch.float16
+dtype = torch.bfloat16
+dtype_vae = torch.bfloat16
+dtype_unet = torch.bfloat16
 unet_needs_upcast = False
